Feat_Front/ClipboardFunction [ADD ClipboardFunction]
